backend/app/core/gemini.py
import google.generativeai as genai
from app.config import settings
from typing import Optional
import logging

logger = logging.getLogger(__name__)

# Configure Gemini API
genai.configure(api_key=settings.GEMINI_API_KEY)


class GeminiAI:
    """Google Gemini AI handler for fashion generation"""

    def __init__(self):
        self.model = genai.GenerativeModel(settings.GEMINI_MODEL)

    async def generate_fashion(self, prompt: str, user_image_url: Optional[str] = None) -> Optional[str]:
        """
        Generate fashion image based on prompt
        Returns: Generated image URL or None
        """
        try:
            # Build the generation prompt
            full_prompt = f"Fashion design generation: {prompt}. Create a realistic fashion item or outfit."

            # For MVP, we'll use text generation
            # In production, you would use Gemini's image generation capabilities
            response = self.model.generate_content(full_prompt)

            # TODO: Implement actual image generation
            # For now, return a placeholder
            logger.info(f"Generated fashion for prompt: {prompt}")
            logger.info(f"Response: {response.text}")

            # In production, you would:
            # 1. Use Gemini's image generation
            # 2. Save the image to local storage
            # 3. Return the image URL
            return f"/uploads/generations/generated_{hash(prompt)}.jpg"  # Placeholder

        except Exception as e:
            logger.error(f"Gemini generation error: {e}")
            return None

    async def try_on_fashion(self, product_image_url: str, user_image_url: str) -> Optional[str]:
        """
        Try on fashion item on user image
        Returns: Generated try-on image URL or None
        """
        try:
            # Build the try-on prompt
            prompt = f"Fashion try-on: Place the clothing item from {product_image_url} onto the person in {user_image_url}. Make it look realistic."

            # For MVP, we'll use text generation
            response = self.model.generate_content(prompt)

            # TODO: Implement actual image generation with try-on
            logger.info(f"Generated try-on for product: {product_image_url}")
            logger.info(f"Response: {response.text}")

            # In production, you would:
            # 1. Load both images
            # 2. Use Gemini's image editing/generation
            # 3. Save the result to local storage
            # 4. Return the image URL
            return f"/uploads/generations/tryon_{hash(product_image_url + user_image_url)}.jpg"  # Placeholder

        except Exception as e:
            logger.error(f"Gemini try-on error: {e}")
            return None


gemini_ai = GeminiAI()
