backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# Database
sqlalchemy==2.0.35
aiosqlite==0.20.0
alembic==1.13.3
greenlet==3.1.1

# Pydantic and validation
pydantic==2.9.2
pydantic-settings==2.5.2
email-validator==2.2.0

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0

# HTTP client
httpx==0.27.2

# PayPal
paypalrestsdk==1.13.3
paypal-checkout-serversdk==1.0.1

# Google Gemini AI
google-generativeai==0.8.3

# Background tasks
apscheduler==3.10.4

# Rate limiting
slowapi==0.1.9

# Email
aiosmtplib==3.0.2
jinja2==3.1.4

# Environment variables
python-dotenv==1.0.1

# Utils
python-dateutil==2.9.0
